Refactoring project before setting up docs
