+++ .gitignore
@@ -3,6 +3,8 @@
 ##
 ## Get latest from https://github.com/github/gitignore/blob/main/VisualStudio.gitignore
 
+*.egg-info/
+
 # User-specific files
 *.rsuser
 *.suo
