+++ app/preparer_requetes.py
