+++ pyproject.toml
@@ -1,0 +1,36 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "stats_dp"
+version = "0.0.1"
+description = "Differentially Private Statistics"
+authors = [
+    { name = "Stuart BENOLIEL", email = "stuart.benoliel@insee.fr" },
+    { name = "Gaspard FEREY", email = "gaspard.ferey@insee.fr" },
+    { name = "Julien JAMME", email = "julien.jamme@insee.fr" },
+]
+readme = "README.md"
+requires-python = ">=3.9"
+dependencies = [
+    "opendp[polars]==0.13.0",
+    "shiny==1.4.0",
+    "shinywidgets==0.6.2",
+    "openpyxl==3.1.5",
+    "cvxpy==1.6.6",
+    "seaborn==0.13.2",
+    "polars==1.12.0",
+    "plotly==6.2.0",
+    "pandas==2.3.0",
+    "PyYAML==6.0.2",
+    "openpyxl==3.1.5",
+    "numpy==2.3.1",
+    "matplotlib==3.10.3"
+]
+
+[tool.setuptools]
+package-dir = {"" = "src"}
+
+[tool.setuptools.packages.find]
+where = ["src"]
