+++ src/stats_dp/constant.py
@@ -1,11 +1,11 @@
-from src.request_class import Query, Count, Sum, Mean, Ratio, Quantile
+from stats_dp.request_class import Query, Count, Sum, Mean, Ratio, Quantile
 import os
 
 # Stockage des options S3 (MinIO)
 storage_options: dict[str, str] = {
-    "aws_access_key_id": os.environ["AWS_ACCESS_KEY_ID"],
-    "aws_secret_access_key": os.environ["AWS_SECRET_ACCESS_KEY"],
-    "aws_session_token": os.environ["AWS_SESSION_TOKEN"],
+#    "aws_access_key_id": os.environ["AWS_ACCESS_KEY_ID"],
+#    "aws_secret_access_key": os.environ["AWS_SECRET_ACCESS_KEY"],
+#    "aws_session_token": os.environ["AWS_SESSION_TOKEN"],
     "endpoint_url": "https://minio.lab.sspcloud.fr",
 }
